+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 Notable changes to this project.
 
 ## [Development]
-- read flag from environment variable to indicate Numerai compute uploads
+- get Numerai compute id if available and pass it along during predictions upload
 
 ## [2.3.2] - 2020-07-21
 - Signals: added `ticker_universe` to get the list of accepted tickers